rename internal_error.rb to server_error.rb
