+++ lib/slack-ruby-client.rb
@@ -28,7 +28,7 @@ require_relative 'slack/web/api/errors/slack_error'
 require_relative 'slack/web/api/errors/too_many_requests_error'
 require_relative 'slack/web/api/error'
 require_relative 'slack/web/api/errors'
-require_relative 'slack/web/api/errors/internal_error'
+require_relative 'slack/web/api/errors/server_error'
 require_relative 'slack/web/faraday/response/raise_error'
 require_relative 'slack/web/faraday/response/wrap_error'
 require_relative 'slack/web/faraday/connection'
