+++ lib/slack/web/api/errors/server_error.rb
